requirements.txt
asgiref==3.2.8
bandit==1.6.2
certifi==2020.4.5.2
chardet==3.0.4
coverage==5.1
Django==3.0.7
entrypoints==0.3
factory-boy==2.12.0
Faker==4.1.1
flake8==3.8.3
flake8-bandit==2.1.2
flake8-builtins==1.5.3
flake8-debugger==3.2.1
flake8-deprecated==1.3
flake8-isort==4.0.0
flake8-logging-format==0.6.0
flake8-polyfill==1.0.2
flake8-print==3.1.4
flake8-quotes==3.2.0
flake8-tuple==0.4.1
gitdb==4.0.5
gitdb2==4.0.2
GitPython==3.1.3
idna==2.9
importlib-metadata==1.6.1
isort==5.3.2
mccabe==0.6.1
natsort==7.0.1
pbr==5.4.5
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
six==1.15.0
smmap==3.0.4
smmap2==3.0.1
sqlparse==0.3.1
stevedore==2.0.0
testfixtures==6.14.1
text-unidecode==1.3
typing==3.7.4.1
unittest-xml-reporting==3.0.2
urllib3==1.25.9
zipp==3.1.0
